Move Document structs to lib
